Use unit test for stage CRs
