+++ kustomize/stage/node/chaos/testdata/node-not-ready.input.yaml
@@ -1,4 +1,3 @@
-# @Stage: ../node-not-ready.yaml
 apiVersion: v1
 kind: Node
 metadata:
