+++ kustomize/stage/node/fast/testdata/node-pending.input.yaml
@@ -1,4 +1,3 @@
-# @Stage: ../node-initialize.yaml
 apiVersion: v1
 kind: Node
 metadata:
