+++ kustomize/stage/pod/fast/testdata/pod-running-with-job.input.yaml
@@ -1,6 +1,3 @@
-# @Stage: ../pod-ready.yaml
-# @Stage: ../pod-delete.yaml
-# @Stage: ../pod-complete.yaml
 apiVersion: v1
 kind: Pod
 metadata:
